# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compliance/check:
    post:
      operationId: runComplianceCheck
      tags: [compliance]
      summary: Run a compliance check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckBody"
      responses:
        "200":
          description: Compliance check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheckResult"
  /compliance/scan-url:
    post:
      operationId: scanUrl
      tags: [compliance]
      summary: Scan a URL for compliance violations (landing page demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanUrlBody"
      responses:
        "200":
          description: URL scan result with violations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanUrlResult"
...
components:
  schemas:
...
    ComplianceCheckBody:
      type: object
      properties:
        module:
          type: string
          description: "Module type: casl, pipeda, bill96, employment, wsib, payroll"
        province:
          type: string
        businessType:
          type: string
        data:
          type: object
          additionalProperties: true
      required:
        - module
        - province
    ComplianceCheckResult:
      type: object
      properties:
        status:
          type: string
          enum: [pass, fail, flag, block]
        score:
          type: number
        title:
          type: string
        statute:
          type: string
        remediation:
          type: string
        module:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - status
        - score
        - title
        - statute
        - remediation
        - module
        - timestamp
    ScanUrlBody:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    ScanUrlResult:
      type: object
      properties:
        url:
          type: string
        overallScore:
          type: number
        violations:
          type: array
          items:
            $ref: "#/components/schemas/ScanViolation"
      required:
        - url
        - overallScore
        - violations
    ScanViolation:
      type: object
      properties:
        law:
          type: string
        issue:
          type: string
        severity:
          type: string
          enum: [high, medium, low]
        citation:
          type: string
      required:
        - law
        - issue
        - severity
        - citation
...