# lib/api-zod/src/generated/types/complianceCheckBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComplianceCheckBodyData } from "./complianceCheckBodyData";

export interface ComplianceCheckBody {
  /** Module type: casl, pipeda, bill96, employment, wsib, payroll */
  module: string;
  province: string;
  businessType?: string;
  data?: ComplianceCheckBodyData;
}