# lib/api-zod/src/generated/types/complianceCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComplianceCheckResultStatus } from "./complianceCheckResultStatus";

export interface ComplianceCheckResult {
  status: ComplianceCheckResultStatus;
  score: number;
  title: string;
  statute: string;
  remediation: string;
  module: string;
  timestamp: Date;
}