# lib/api-zod/src/generated/types/scanUrlResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanViolation } from "./scanViolation";

export interface ScanUrlResult {
  url: string;
  overallScore: number;
  violations: ScanViolation[];
}