# lib/api-zod/src/generated/types/scanViolation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanViolationSeverity } from "./scanViolationSeverity";

export interface ScanViolation {
  law: string;
  issue: string;
  severity: ScanViolationSeverity;
  citation: string;
}